requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
aiohttp==3.13.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
aiosignal==1.4.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
alembic==1.17.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
anyio==4.12.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
asyncer==0.0.8 ; python_version >= "3.11" and python_full_version <= "3.13.0"
attrs==25.4.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
backoff==2.2.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
cachetools==6.2.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
certifi==2025.11.12 ; python_version >= "3.11" and python_full_version <= "3.13.0"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
click==8.3.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
cloudpickle==3.1.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
colorama==0.4.6 ; python_version >= "3.11" and python_full_version <= "3.13.0" and (platform_system == "Windows" or sys_platform == "win32")
colorlog==6.10.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
diskcache==5.6.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
distro==1.9.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
dspy==3.0.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
fastuuid==0.14.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
filelock==3.20.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
frozenlist==1.8.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
fsspec==2025.12.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
gepa[dspy]==0.0.17 ; python_version >= "3.11" and python_full_version <= "3.13.0"
greenlet==3.2.4 ; python_version >= "3.11" and python_full_version <= "3.13.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
grpcio==1.67.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
h11==0.16.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
hf-xet==1.2.0 ; python_version >= "3.11" and python_full_version <= "3.13.0" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "arm64" or platform_machine == "aarch64")
httpcore==1.0.9 ; python_version >= "3.11" and python_full_version <= "3.13.0"
httpx==0.28.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
huggingface-hub==1.1.7 ; python_version >= "3.11" and python_full_version <= "3.13.0"
idna==3.11 ; python_version >= "3.11" and python_full_version <= "3.13.0"
importlib-metadata==8.7.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_full_version <= "3.13.0"
jiter==0.12.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
joblib==1.5.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
json-repair==0.54.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
jsonschema-specifications==2025.9.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
jsonschema==4.25.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
litellm==1.80.7 ; python_version >= "3.11" and python_full_version <= "3.13.0"
magicattr==0.1.6 ; python_version >= "3.11" and python_full_version <= "3.13.0"
mako==1.3.10 ; python_version >= "3.11" and python_full_version <= "3.13.0"
markdown-it-py==4.0.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
markupsafe==3.0.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
multidict==6.7.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
numpy==2.3.5 ; python_version >= "3.11" and python_full_version <= "3.13.0"
openai==2.8.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
optuna==4.6.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
orjson==3.11.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
packaging==25.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pillow==12.0.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
propcache==0.4.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pydantic-core==2.41.5 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pydantic==2.12.5 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pygments==2.19.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
python-dotenv==1.2.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
pyyaml==6.0.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
referencing==0.37.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
regex==2025.11.3 ; python_version >= "3.11" and python_full_version <= "3.13.0"
requests==2.32.5 ; python_version >= "3.11" and python_full_version <= "3.13.0"
rich==14.2.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
rpds-py==0.30.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_full_version <= "3.13.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
sqlalchemy==2.0.44 ; python_version >= "3.11" and python_full_version <= "3.13.0"
tenacity==9.1.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
tiktoken==0.12.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
tokenizers==0.22.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_full_version <= "3.13.0"
typer-slim==0.20.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_full_version <= "3.13.0"
urllib3==2.5.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
xxhash==3.6.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
yarl==1.22.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
zipp==3.23.0 ; python_version >= "3.11" and python_full_version <= "3.13.0"
